+++ scripts/es2k/report-krnlmon-coverage.sh
@@ -31,7 +31,7 @@ TEST_LABEL=krnlmon
 mkdir -p ${OUTPUT_DIR}
 
 # Empty it for good measure.
-rm -fr ${OUTPUT_DIR}/*
+rm -fr ${OUTPUT_DIR:?}/*
 
 # Capture the coverage data, excluding files in directories
 # that don't interest us.
